Separate fuzz tests by Go version
